requirements.txt
aws==0.2.5
awscli==1.16.309
bcrypt==3.1.7
boto==2.49.0
boto3==1.10.45
botocore==1.13.45
cffi==1.13.2
colorama==0.4.1
cryptography==2.8
dj-database-url==0.5.0
Django==2.2.6
django-heroku==0.3.1
django-resized==0.3.9
django-storages==1.8
docutils==0.15.2
fabric==2.5.0
gunicorn==20.0.0
invoke==1.3.0
jmespath==0.9.4
paramiko==2.7.1
Pillow==6.2.1
prettytable==0.7.2
psycopg2==2.8.4
pyasn1==0.4.8
pycparser==2.19
PyNaCl==1.3.0
python-dateutil==2.8.1
python-dotenv==0.10.3
pytz==2019.3
PyYAML==5.2
rsa==3.4.2
s3transfer==0.2.1
six==1.13.0
sqlparse==0.3.0
urllib3==1.25.7
whitenoise==4.1.4
